requirements.txt
anyio==3.7.1
asgiref==3.7.2
certifi==2023.7.22
Django==3.2.20
exceptiongroup==1.1.2
gunicorn==21.2.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-metadata==6.7.0
packaging==23.1
psycopg2==2.9.9
python-telegram-bot==20.2
pytz==2023.3
rfc3986==1.5.0
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.7.1
whitenoise==6.5.0
zipp==3.15.0
